requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
simpy>=4.0.1
python-dotenv>=0.19.0

# Optional dependencies for advanced features
scikit-learn>=1.0.0  # For more advanced analysis
networkx>=2.6.0      # For agent network visualization
seaborn>=0.11.0      # For enhanced visualizations
ipython>=7.0.0       # For interactive analysis

# Web dashboard
Flask>=2.0.0

# LLM integration
google-generativeai>=0.3.0  # For Gemini API integration
openai>=1.0.0  # For OpenAI GPT integration
